Extract component DialogFormField
